+++ kernel/include/b1nix/rtc.h
@@ -4,9 +4,9 @@
 #include <b1nix/types.h>
 
 void rtc_init(void);
-u32 rtc_now_unix_seconds(void);
-void rtc_set_unix_time(u32 unix_time_now);
+u64 rtc_now_unix_seconds(void);
+void rtc_set_unix_time(u64 unix_time_now);
 
-extern u32 rtc_boot_time_seconds;
+extern u64 rtc_boot_time_seconds;
 
 #endif
