+++ userspace/include/sys/types.h
@@ -8,7 +8,7 @@ typedef long ssize_t;
 typedef long off_t;
 #ifndef B1NIX_TIME_T_DEFINED
 #define B1NIX_TIME_T_DEFINED
-typedef long time_t;
+typedef long long time_t;
 #endif
 typedef int pid_t;
 typedef unsigned int mode_t;
